+++ src/game/player.rs
@@ -7,11 +7,9 @@ use bevy_aseprite_ultra::prelude::*;
 use bevy_ecs_ldtk::prelude::*;
 use bevy_rapier2d::prelude::*;
 use std::f32::consts::PI;
-use std::fmt::Display;
-use std::fmt::Formatter;
 
 #[derive(Component)]
-pub struct Person;
+pub struct Player;
 
 pub fn setup_player(
     mut commands: Commands,
@@ -19,7 +17,7 @@ pub fn setup_player(
     player_data: &PlayerData,
 ) {
     commands.spawn((
-        Person,
+        Player,
         AsepriteAnimationBundle {
             aseprite: asset_server.load("player.aseprite"),
             transform: Transform::from_xyz(player_data.x, player_data.y, 1.0),
@@ -50,11 +48,11 @@ pub fn update_player(
             &GlobalTransform,
             &mut Sprite,
         ),
-        (With<Person>, Without<Camera2d>),
+        (With<Player>, Without<Camera2d>),
     >,
     mut camera_query: Query<
         (&Camera, &mut Transform, &GlobalTransform),
-        (With<Camera2d>, Without<Person>),
+        (With<Camera2d>, Without<Player>),
     >,
 
     ldtk_projects: Query<&Handle<LdtkProject>>,
@@ -84,39 +82,44 @@ pub fn update_player(
     //     player.translation.y, player.translation.z
     // );
 
-    let (camera, mut camera_transform, camera_global_transform) = camera_query.single_mut();
+    if let Ok((camera, mut camera_transform, camera_global_transform)) =
+        camera_query.get_single_mut()
+    {
+        camera_transform.translation.x +=
+            (player.translation.x - camera_transform.translation.x) * 0.1;
+        camera_transform.translation.y +=
+            (player.translation.y - camera_transform.translation.y) * 0.1;
 
-    camera_transform.translation.x += (player.translation.x - camera_transform.translation.x) * 0.1;
-    camera_transform.translation.y += (player.translation.y - camera_transform.translation.y) * 0.1;
+        // プレイヤーの向き
+        if let Ok(window) = q_window.get_single() {
+            let cursor = window
+                .cursor_position()
+                .and_then(|cursor| camera.viewport_to_world(camera_global_transform, cursor))
+                .map(|ray| ray.origin.truncate())
+                .unwrap_or(Vec2::ZERO);
 
-    // プレイヤーの向き
-    let window = q_window.single();
-    let cursor = window
-        .cursor_position()
-        .and_then(|cursor| camera.viewport_to_world(camera_global_transform, cursor))
-        .map(|ray| ray.origin.truncate())
-        .unwrap_or(Vec2::ZERO);
+            let ray = cursor - player.translation.truncate();
+            let angle = ray.y.atan2(ray.x);
+            // println!("angle: {}", angle);
+            if angle < -PI * 0.5 || PI * 0.5 < angle {
+                player_sprite.flip_x = true;
+            } else {
+                player_sprite.flip_x = false;
+            }
 
-    let ray = cursor - player.translation.truncate();
-    let angle = ray.y.atan2(ray.x);
-    // println!("angle: {}", angle);
-    if angle < -PI * 0.5 || PI * 0.5 < angle {
-        player_sprite.flip_x = true;
-    } else {
-        player_sprite.flip_x = false;
-    }
-
-    // レベルの追従
-    // https://trouv.github.io/bevy_ecs_ldtk/v0.10.0/how-to-guides/make-level-selection-follow-player.html
-    // ただし上記の方法では読み込み済みのレベルを利用して現在のレベルを判定しているため、
-    // 歩いて移動する場合は機能しますが、プレイヤーがワープしたり、セーブした位置から再開する場合は機能しません。
-    // このため、常にアセット全体を検索して現在のレベルを判定しています
-    let player_position = Vec2::new(player.translation.x, player.translation.y);
-    if let Ok(project_handle) = ldtk_projects.get_single() {
-        if let Some(ldtk_project) = ldtk_project_assets.get(project_handle) {
-            let found = find_level(ldtk_project, player_position);
-            if let Some(level_iid) = found {
-                *level_selection = LevelSelection::Iid(level_iid.clone());
+            // レベルの追従
+            // https://trouv.github.io/bevy_ecs_ldtk/v0.10.0/how-to-guides/make-level-selection-follow-player.html
+            // ただし上記の方法では読み込み済みのレベルを利用して現在のレベルを判定しているため、
+            // 歩いて移動する場合は機能しますが、プレイヤーがワープしたり、セーブした位置から再開する場合は機能しません。
+            // このため、常にアセット全体を検索して現在のレベルを判定しています
+            let player_position = Vec2::new(player.translation.x, player.translation.y);
+            if let Ok(project_handle) = ldtk_projects.get_single() {
+                if let Some(ldtk_project) = ldtk_project_assets.get(project_handle) {
+                    let found = find_level(ldtk_project, player_position);
+                    if let Some(level_iid) = found {
+                        *level_selection = LevelSelection::Iid(level_iid.clone());
+                    }
+                }
             }
         }
     }
@@ -124,34 +127,4 @@ pub fn update_player(
 
 fn to_s(keys: &Res<ButtonInput<KeyCode>>, code: bevy::input::keyboard::KeyCode) -> f32 {
     return if keys.pressed(code) { 1.0 } else { 0.0 };
-}
-
-pub enum PlayerDirection {
-    PlayerUp,
-    PlayerDown,
-    PlayerLeft,
-    PlayerRight,
-}
-
-impl Display for PlayerDirection {
-    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
-        match self {
-            PlayerDirection::PlayerUp => write!(f, "PlayerUp"),
-            PlayerDirection::PlayerDown => write!(f, "PlayerDown"),
-            PlayerDirection::PlayerLeft => write!(f, "PlayerLeft"),
-            PlayerDirection::PlayerRight => write!(f, "PlayerRight"),
-        }
-    }
-}
-
-pub fn angle_to_direction(angle: f32) -> PlayerDirection {
-    if PI * -0.75 <= angle && angle < PI * -0.25 {
-        return PlayerDirection::PlayerDown;
-    } else if PI * -0.25 <= angle && angle < 0.25 {
-        return PlayerDirection::PlayerRight;
-    } else if 0.25 <= angle && angle < PI * 0.75 {
-        return PlayerDirection::PlayerUp;
-    } else {
-        return PlayerDirection::PlayerLeft;
-    }
 }
